Collection Script Prototype
Collection script for the Arduino is complete. Through one single program, data can be collected and predictions can be done. The predictions call the MLModel script and runs a simple prediction based on whatever model was run on it before.
Checked for most edge cases when using the program. There is sometime an issue with the serial communication between the Arduino and Python which I currently handle with an error warning. Restarting the kernel(program) helps to resolve this issue.

TODO: Collect more data to build out more ML models in MLModels.
Test to find out what happens when the data read from the queue fails.
In MLModels: Impute null values.